Use relative instead of absolute time inside solvers (#44)
The solvers now integrate from t0 to t0 + Δt_cpl,
instead of knowing internally about t_max (total duration of simulation)
Main effect: CIT boundary mapping now works even when t_max > Δt_cpl (i.e., #42).
Also, the checkpoint files are now overwritten in each new coupling window, leading to much less files generated during a simulation.
While I was at it, I refactored the atm/oce RHS functions to reduce code duplication.

+++ src/parameters.jl
@@ -34,8 +34,8 @@ Base.@kwdef mutable struct SimulationParameters
     T_O_ini = 271.0
     T_I_ini = 270.0
     h_I_ini = 1.0
-    t_max = 3600.0
-    Δt_cpl = 100.0
+    t_max::Float64 = 3600.0
+    Δt_cpl::Float64 = 100.0
     Δt_min = 1.0
     n_t_A = 50
     n_t_O = 1
